feat: add test for lotto numbers
