+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -22,6 +22,4 @@ class WinningNumbersTest {
 				//then
 				.isInstanceOf(LottoException.class);
 	}
-
-
 }
