+++ src/test/java/lotto/domain/vo/LottoNumbersTest.java
@@ -1,30 +1,11 @@
 package lotto.domain.vo;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-
 import java.util.List;
-import lotto.exception.LottoException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class LottoNumbersTest {
-
-
-	@DisplayName("당첨 번호가 6개 이하나 이상이라면 예외가 발생한다.")
-	@Test
-	void underSizeWinnings() {
-		//given
-		List<Integer> underSizeNumber = List.of(1, 2, 3, 4, 5);
-		List<Integer> overSizeNumber = List.of(1, 2, 3, 4, 5, 6, 7);
-		//when & then
-		//under
-		assertThatThrownBy(() -> new LottoNumbers(underSizeNumber))
-				.isInstanceOf(LottoException.class);
-		//over
-		assertThatThrownBy(() -> new LottoNumbers(overSizeNumber))
-				.isInstanceOf(LottoException.class);
-	}
 
 	@DisplayName("로또 숫자에 일치하는 숫자가 포함여부에 따라 true와 false를 반환한다.")
 	@Test
